pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vector"
version = "0.1.0"
description = "Issue-conditioned discovery & ranking system"
authors = [{name = "Parallel LLC", email = "engineering@parallel.example"}]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.0.0",
  "numpy>=1.23.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0",
  "typer>=0.12.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
  "scikit-learn>=1.3.0",
  "networkx>=3.2.0"
]

[project.scripts]
vector = "vector.cli:app"
